+++ SMPHelper.h
@@ -12,14 +12,14 @@ public:
 	SMPHelper();
 	~SMPHelper();
 	//Wygeneruj macierz srednich dla A i B
-	boost::numeric::ublas::matrix<double> GenerateAvarageMatrixForFeatures
-	(std::vector<int> featureComb, std::map<int, double> avarages, int objCount, int dimension);
+	boost::numeric::ublas::matrix<long double> GenerateAvarageMatrixForFeatures
+	(std::vector<int> featureComb, std::map<int, long double> avarages, int objCount, int dimension);
 
 	void GenerateXMatrixForFeatures
 	(std::vector<int> featureComb, Database database, std::map<std::string, int> objectCount,
-		std::vector<std::string> classNames, boost::numeric::ublas::matrix<double>& MatrixXa, boost::numeric::ublas::matrix<double>& MatrixXb);
+		std::vector<std::string> classNames, boost::numeric::ublas::matrix<long double>& MatrixXa, boost::numeric::ublas::matrix<long double>& MatrixXb);
 
-	double CalculateUa_UbAvaragesLength
-	(std::vector<int> featureComb, std::map<int, double> avaragesA, std::map<int, double> avaragesB);
+	long double CalculateUa_UbAvaragesLength
+	(std::vector<int> featureComb, std::map<int, long double> avaragesA, std::map<int, long double> avaragesB);
 };
